user-settings-api: implement-api

+++ server/services/user.service.ts
@@ -24,7 +24,13 @@ export const updateById = async (id: string, body: any): Promise<IResponse> => {
   return await getCustomRepository(UserRepository).updateById(id, body);
 };
 
-export const deleteById = async (id: string): Promise<IResponse> => {
+export const deleteById = async (
+  id: string,
+  user: User,
+  next
+): Promise<IResponse> => {
+  if (id !== user.id)
+    return next({ status: 401, message: "Permision denied." }, null);
   return await getCustomRepository(UserRepository).deleteById(id);
 };
 
@@ -36,7 +42,14 @@ export const getByUserName = async (name): Promise<User> => {
   return await getCustomRepository(UserRepository).findOne({ name });
 };
 
-export const updateEmail = async (user: User, email: string, next) => {
+export const updateEmail = async (
+  userId: string,
+  user: User,
+  email: string,
+  next
+) => {
+  if (userId !== user.id)
+    return next({ status: 401, message: "Permision denied." }, null);
   const userByEmail = await getByEmail(email);
   if (userByEmail) {
     return next({ status: 401, message: "Email is already taken." }, null);
@@ -47,7 +60,14 @@ export const updateEmail = async (user: User, email: string, next) => {
   });
 };
 
-export const updatePassword = async (user: User, password: string, next) => {
+export const updatePassword = async (
+  userId: string,
+  user: User,
+  password: string,
+  next
+) => {
+  if (userId !== user.id)
+    return next({ status: 401, message: "Permision denied." }, null);
   if (password.length < 6) {
     return next(
       { status: 401, message: "Password shoud contain min 6 symbols" },
@@ -57,5 +77,58 @@ export const updatePassword = async (user: User, password: string, next) => {
   return await getCustomRepository(UserRepository).updateById(user.id, {
     password,
     reset_token: ""
+  });
+};
+
+interface IUserNotificationSettings {
+  emailNotificationNews?: boolean;
+  emailNotificationUpdatesFromFollowed?: boolean;
+  emailNotificationComments?: boolean;
+  emailNotificationEvents?: boolean;
+  siteNotificationUpdatesFromFollowed?: boolean;
+  siteNotificationComments?: boolean;
+  siteNotificationEvents?: boolean;
+}
+
+export const updateNotificationSettings = async (
+  userId: string,
+  user: User,
+  newSettings: IUserNotificationSettings,
+  next
+) => {
+  if (userId !== user.id)
+    return next({ status: 401, message: "Permision denied." }, null);
+  return await getCustomRepository(UserRepository).updateById(userId, {
+    ...user,
+    ...newSettings
+  });
+};
+
+type Privacy = "All" | "Followers" | "OnlyMe";
+
+interface IUserPrivacySettings {
+  privacyProfileInfo?: Privacy;
+  privacyMyPosts?: Privacy;
+  privacyStories?: Privacy;
+  privacyEvents?: Privacy;
+  privacySurveys?: Privacy;
+  privacyTops?: Privacy;
+  privacyCollections?: Privacy;
+  privacyWatchlist?: Privacy;
+  privacyReviews?: Privacy;
+  privacyMessages?: Privacy;
+}
+
+export const updatePrivacySettings = async (
+  userId: string,
+  user: User,
+  newSettings: IUserPrivacySettings,
+  next
+) => {
+  if (userId !== user.id)
+    return next({ status: 401, message: "Permision denied." }, null);
+  return await getCustomRepository(UserRepository).updateById(userId, {
+    ...user,
+    ...newSettings
   });
 };
